Move version to separate file in order to not cycle deps
